Rename and refactor project (#23)
