+++ apps/apple/VulkanSplatting/VulkanSplatting-Bridging-Header.h
@@ -2,7 +2,7 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#include "VulkanSplatting.h"
+#include "3dgs.h"
 #include <memory>
 #include <iostream>
